+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
-* (22 May 2017) -- Initial Release
+## [1.0.2] - 2017-05-30
+###Added
+- Removal of emojis and generator tags
+- Sets default WYSIWYG editor to HTML instead
+## [1.0.1] - 2017-05-22
+- Initial release